Fix some threading issues:
- QThread::terminate() is too heavy-impact for our purposes.
- There were two concurrent ending paths: 1. thread termination via
  QThread::terminate(), 2. run end via QThread::run().

As a consequence, dact deadlocked when starting a new query when an
existing query was running.

Rather than providing terminate, we now provide a cancel() method that
flags that the thread should be canceled. This ends the main loop in
the run method().

Todo: when a new thread is started, there may still be unprocessed signals
from the previous query.