Binary search in QueryModel::mapperEntryFound.
This is a partial fix for #24. Observe that, in order to do a binary
search the list should be ordered on all members of the entry (and not
just the number of hits). Also, you will want to perform the binary
search using the old item (before updating the hit count). Since changing
the hit count before removing the affected item from the list, makes the
list unordered.