Quickly added an option to export the raw XML
I needed this to fiddle with the stylesheets outside of Dact as I don't have the original xml files laying around anymore. I think it is a usefull function for others as well.

It doesn't do multiple files yet, but neither does exportPDF. Might be usefull, for both, to add this later on?