Fire query when user selects a filter query from the dropdown menu.
This listens for click-signals on the dropdown list view and sets a flag if one occurred. When the selection changes it will fire the enterOrClick signal if the flag is true. This way, only when the user used the mouse to select an alternative query the event is fired, not when the user used cursor keys. This sort-of fixes issue #75.