Reduce lock contention in FilterModel.
In FilterModel, a lot of locking is used to ensure the result list is
always observed in a valid state. However, the list is locked very often
for updates, blocking all reads that update the user interface.

So, instead, we maintain two lists: one that that backs the actual model and
one that is rapidly updated through query iteration. The list backing the
model uses a QReadWriteLock, to avoid that multiple reads are blocking
eachother. A write to that list only occurs in fireDateChanged(), to replace
the model-backing list by the updated list.

When profiling, this removes the lock overhead. This is a part of the
solution to #123. We are not there yet, since the interface is still slow
due to all the field width calculations that Qt needs to perform.