Downgrade to DB XML 2.5.16.
It seems that Berkeley DB 6.x and Berkeley DB XML 6.x switched to Affero
GPL. This means that we'd have to relicense everything, since the LGPL 2.1
is not compatible with the Affero GPL version 3. Also, using Affero GPL for
a library is such an unholy mess that even Debian does not want to go there.

tl;dr: Oracle, you suck!

Revert "Update AST processing."

This reverts commit 56d0627c0b7b20f0163b97bbc3b091488a4f8dca.