Add parenthesis around predicates before query expansion.
The query expansion gives incorrect results when the predicate
is a disjunction (#132). E.g.:

//node[@word="bezorgd" or @word="bezorgde"]

was incorrectly expanded to:

//node[@word="bezorgd" or @word="bezorgde" and CONDITION]

rather than the correct

//node[(@word="bezorgd" or @word="bezorgde") and CONDITION]

Ideally we should do this using AST manipulation. Though it
seems that XQilla does not provide any obvious way to serialize
an expression AST to a query string. So, we proceed doing this
the ugly way.

Fixes #132.