Add Qt5 to prefix path on keg-ized Qt5.
Qt5 is installed as a keg by Homebrew on macOS. Homebrew adds Qt to
Homebrew's prefix path if Dact is installed through Homebrew.  However,
it is (obviously) not in manual builds. So, we add /usr/local/opt/qt5 to
the CMake prefix path if were are on an Apple platform and that path
exists.